[value-path-wiring] wire custody signing + real broadcast + ulong amounts + KYC choke point
Closes the review's Part-A value-flow gaps so the path actually moves value
correctly:
- C1: route Algorand signing through IKeyCustodyService via a SigningContext
  {AvatarId,WalletId,IsPlatform} — per-user ops sign with the user's key
  (WithSigningKeyAsync, IDOR-guarded), platform/ASA-admin ops use
  WithPlatformSigningKeyAsync; a user op with no resolvable context FAILS CLOSED
  (no silent platform mis-sign). Scoped custody resolved per-op via
  IServiceScopeFactory (singleton provider, mirrors AlgorandFaucet).
- C2: AllocationManager builds a typed op and drives IBlockchainOperationManager
  .ExecuteAsync so it broadcasts + records a real TxHash; the idempotency key is
  Completed only with a TxHash (else InProgress).
- H1: the alloc:{apiKeyId} key is persisted on the op row for crash recovery.
- H3: the KYC gate moved into NftManager.MintAsync (single choke point) so raw
  POST /api/nft/mint is gated too; KYC_FORBIDDEN -> 403.
- H4: amount widened int->ulong end-to-end (IMintOperation/BlockchainOperation/
  provider surface) — no clamp, the idempotency key keys off the true ulong.
- M1: confirm-timeout records PendingConfirmation + TxHash so reconciliation
  settles from chain truth and never re-broadcasts.
Narrowing casts removed; the two remaining (CreateASA total, sim ledger) are
checked. Crypto independently reviewed (separate lane).

+++ Controllers/HolonController.cs
@@ -202,7 +202,7 @@ public class MintRequest
 {
     public Guid WalletId { get; set; }
     public string TokenUri { get; set; } = string.Empty;
-    public int Amount { get; set; }
+    public ulong Amount { get; set; }
     public string AssetType { get; set; } = string.Empty;
 }
