+++ Interfaces/IMintOperation.cs
@@ -3,6 +3,6 @@ namespace OASIS.WebAPI.Interfaces;
 public interface IMintOperation : IBlockchainOperation
 {
     string? TokenUri { get; set; }
-    int Amount { get; set; }
+    ulong Amount { get; set; }
     string? AssetType { get; set; }
 }
